code refactored, cargo fmt done, 'json' name folder change to 'upload_json' and embed for upload_json response
