+++ src/commands/duo_stats/mod.rs
@@ -1,3 +1,3 @@
 pub mod get_duo_stats;
-pub mod utils;
 pub mod modal;
+pub mod utils;
