+++ src/commands/mob_stats/mod.rs
@@ -1,3 +1,3 @@
 pub mod get_mob_stats;
-pub mod utils;
 pub mod modal;
+pub mod utils;
