+++ src/commands/mod.rs
@@ -1,6 +1,6 @@
 pub mod duo_stats;
 pub mod help;
-pub mod json;
+pub mod upload_json;
 pub mod mob_stats;
 pub mod player_names;
 pub mod ranks;
