+++ src/commands/player_names/mod.rs
@@ -1,4 +1,4 @@
-pub mod track_player_names;
-pub mod utils;
 pub mod modal;
 pub mod models;
+pub mod track_player_names;
+pub mod utils;
