+++ src/commands/shared/mod.rs
@@ -1,4 +1,4 @@
-pub mod models;
-pub mod utils;
 pub mod embed_error_handling;
 pub mod logs;
+pub mod models;
+pub mod utils;
