+++ src/commands/shared/models.rs
