+++ src/commands/suggestion/mod.rs
@@ -1,2 +1,2 @@
-pub mod send_suggestion;
 pub mod modal;
+pub mod send_suggestion;
