+++ src/main.rs
@@ -10,7 +10,7 @@ use std::sync::{Arc, Mutex};
 
 use crate::commands::duo_stats::get_duo_stats::get_duo_stats;
 use crate::commands::help::help::help;
-use crate::commands::json::upload_json::upload_json;
+use crate::commands::upload_json::upload_json::upload_json;
 use crate::commands::mob_stats::get_mob_stats::get_mob_stats;
 use crate::commands::player_names::track_player_names::track_player_names;
 use crate::commands::ranks::get_ranks::get_ranks;
